Figured out the problem, don't know best way of fixing it yet:
With each iteration, the OrderedPairs are keeping the RV that was
created in the previous runthrough of findBestRV(). Because of this,
absurdly high rhyme values are being generated for each runthrough of
the findBestRV() method. This ends up creating a sort of parabola
effect though since there is very little or no extra rhyme value being
added to OrderedPairs at the very bottom. So have to figure out an
efficient way of resetting to the original rhyme value for each ordered
pair with each runthrough.